# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TRS/OEE DPI Pharmaceutical tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /planning/parse:
    post:
      operationId: parsePlanningFile
      tags: [planning]
      summary: Parse an Excel planning file and return structured preview
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Parsed planning preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedPlanningResult"
  /planning/import:
    post:
      operationId: importPlanning
      tags: [planning]
      summary: Import parsed planning rows into the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPlanningBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                required: [imported]
  /planning:
    get:
      operationId: listPlanning
      tags: [planning]
      summary: List production plans
      parameters:
        - name: week
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of plan entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanEntry"
...
  /planning/{id}/validate:
    patch:
      operationId: validatePlanEntry
      tags: [planning]
      summary: Validate or reject a plan entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidatePlanBody"
      responses:
        "200":
          description: Updated plan status
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  validationStatus:
                    type: string
                required: [id, validationStatus]

  # Notifications
...
components:
  schemas:
...
    PlanRow:
      type: object
      properties:
        weekNumber:
          type: integer
        year:
          type: integer
        date:
          type: string
        dayName:
          type: string
        activityType:
          type: string
        team:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        room:
          type: ["string", "null"]
        productName:
          type: ["string", "null"]
        lotNumber:
          type: ["string", "null"]
        plannedQuantity:
          type: ["number", "null"]
        plannedUnit:
          type: ["string", "null"]
        specialActivity:
          type: ["string", "null"]
      required: [weekNumber, year, date, dayName, activityType]
    ParsedPlanningResult:
      type: object
      properties:
        weekNumber:
          type: integer
        year:
          type: integer
        fileName:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/PlanRow"
        anomalies:
          type: array
          items:
            type: string
      required: [weekNumber, year, fileName, rows, anomalies]
    ImportPlanningBody:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/PlanRow"
        fileName:
          type: string
      required: [rows, fileName]
    PlanEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        weekNumber:
          type: integer
        year:
          type: integer
        date:
          type: string
        dayName:
          type: string
        activityType:
          type: string
        team:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        room:
          type: ["string", "null"]
        productName:
          type: ["string", "null"]
        lotNumber:
          type: ["string", "null"]
        plannedQuantity:
          type: ["number", "null"]
        plannedUnit:
          type: ["string", "null"]
        specialActivity:
          type: ["string", "null"]
        sourceFileName:
          type: string
        importedAt:
          type: string
          format: date-time
        validationStatus:
          type: string
        importedByName:
          type: ["string", "null"]
      required: [id, weekNumber, year, date, dayName, activityType, sourceFileName, importedAt, validationStatus]
...
    ValidatePlanBody:
      type: object
      properties:
        status:
          type: string
          enum: [validated, rejected]
        comment:
          type: string
      required: [status]

    # Notifications
...